requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cors==1.0.1
Django==3.2.12
django-cors-headers
filelock==3.13.3
future==1.0.0
gevent==24.2.1
greenlet==3.0.3
idna==3.6
pycparser==2.22
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.31.0
requests-file==2.0.0
setuptools==69.2.0
sqlparse==0.4.4
tldextract==5.1.2
tzdata==2024.1
urllib3==2.2.1
zope.event==5.0
zope.interface==6.2
djangorestframework 
Pillow
